Move checkbox group functions to seperate module
