+++ app/questionGroup/get.controller.js
@@ -4,31 +4,7 @@ const {
   compileInlineConditionalQuestions,
   grabAnswers,
 } = require('../../common/question-groups/get-question-groups')
-const questionResponse = require('../../wiremock/responses/questionResponse')
-
-const flattenCheckboxGroups = questions => {
-  return questions.map(question => {
-    if (question.type === 'checkboxGroup') {
-      return {
-        type: 'question',
-        questionId: question.checkboxGroupId,
-        questionCode: question.checkboxGroupCode,
-        answerType: 'checkbox',
-        questionText: question.title,
-        displayOrder: question.displayOrder,
-        mandatory: question.mandatory || true,
-        readOnly: question.readOnly || false,
-        conditional: question.conditional || false,
-        answerSchemas: question.contents.map(({ questionId, questionText }) => ({
-          value: questionId,
-          text: questionText,
-        })),
-      }
-    }
-
-    return question
-  })
-}
+const { flattenCheckboxGroups } = require('../../common/data/checkboxGroups')
 
 const displayQuestionGroup = async (
   { params: { assessmentId, groupId, subgroup }, body, errors = {}, errorSummary = null, user },
@@ -62,4 +38,4 @@ const displayQuestionGroup = async (
   }
 }
 
-module.exports = { displayQuestionGroup, flattenCheckboxGroups }
+module.exports = { displayQuestionGroup }
